Checkout implements User info
If the user is logged in, less of the checkout form is shown. Furthermore, users with an empty basket can no longer access Checkout.php or update.php. Basket.php now detects if the basket is empty. redirect.php will correctly use the UserID or create a guest account in the right situation. The £4.99 cost is added to the total amount for next-day shipping. confirm.php has been converted to .html.